requirements.txt
requests
sqlalchemy
psycopg
python-dotenv